+++ packages/background/package.json
@@ -16,6 +16,9 @@
   "dependencies": {
     "@coral-xyz/common": "*",
     "@coral-xyz/recoil": "*",
+    "@coral-xyz/blockchain-common": "*",
+    "@coral-xyz/blockchain-solana": "*",
+    "@coral-xyz/blockchain-evm": "*",
     "@solana/web3.js": "^1.63.1",
     "bip39": "^3.0.4",
     "bs58": "^5.0.0",
