+++ packages/background/src/backend/core.ts
@@ -13,6 +13,10 @@ import {
 import type { KeyringStoreState } from "@coral-xyz/recoil";
 import { makeDefaultNav } from "@coral-xyz/recoil";
 import {
+  keyringForBlockchain,
+  BlockchainKeyring,
+} from "@coral-xyz/blockchain-common";
+import {
   BACKEND_EVENT,
   Blockchain,
   EthereumConnectionUrl,
@@ -45,7 +49,6 @@ import {
   SolanaExplorer,
   deserializeTransaction,
   FEATURE_GATES_MAP,
-  XnftPreferenceStore,
   XnftPreference,
 } from "@coral-xyz/common";
 import type {
@@ -56,7 +59,6 @@ import type {
 } from "@coral-xyz/common";
 import type { Nav } from "./store";
 import * as store from "./store";
-import { BlockchainKeyring } from "./keyring/blockchain";
 import { KeyringStore } from "./keyring";
 import type { SolanaConnectionBackend } from "./solana-connection";
 import type { EthereumConnectionBackend } from "./ethereum-connection";
@@ -438,10 +440,7 @@ export class Backend {
     publicKey: string,
     mnemonic?: string
   ) {
-    const blockchainKeyring = {
-      [Blockchain.SOLANA]: BlockchainKeyring.solana,
-      [Blockchain.ETHEREUM]: BlockchainKeyring.ethereum,
-    }[blockchain]();
+    const blockchainKeyring = keyringForBlockchain(blockchain);
 
     if (mnemonic) {
       blockchainKeyring.initFromMnemonic(mnemonic, derivationPath, [
