+++ packages/blockchains/common/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@coral-xyz/blockchain-common",
+  "version": "0.1.0",
+  "main": "src/index.js",
+  "license": "MIT"
+}
