+++ packages/blockchains/solana/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@coral-xyz/blockchain-solana",
+  "version": "0.1.0",
+  "main": "src/index.ts",
+  "license": "MIT"
+}
